requirements.txt
pandas
numpy
torch
tensorflow
pyyaml
python-dotenv
matplotlib
scikit-learn
scipy
opencv-python
flask
hop
torchvision
tqdm
ultralytics
seaborn
ipykernel
gdown
Ipython
dvc

-e .